Moved classes into dinojump package and DatabaseManager into the manager package
